+++ includes/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: walnaimi <walnaimi@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: fdessoy- <fdessoy-@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/13 10:13:01 by fdessoy-          #+#    #+#             */
-/*   Updated: 2024/08/14 16:22:02 by walnaimi         ###   ########.fr       */
+/*   Updated: 2024/08/14 14:35:09 by fdessoy-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,9 +91,8 @@ typedef struct s_data
 	int			status;
 	char		**cmd; // need to insert pids, tokens, and commands
 	t_token		*token;
-	t_token		*first_node;
-	t_token		*current_token;
-	t_token		*prev_token;
+	char		**builtins;
+	char		**redirect;
 	char		**cmd_a;
 	bool		echoed;
 	bool		echo_flag;
@@ -113,13 +112,10 @@ typedef struct s_data
 	int			in_quotes;//ft_strtok
 	char		quote_char;//ft_strtok
 	size_t		len_t;
-	size_t		len_env;
-	int			env_copied;
+	size_t		i_t;
+	size_t		j_t;
 	int			s_quote_o;
 	int			d_quote_o;
-	int			exit_code;
-	char		*new_str;
-	char		*temp_str;
 	t_env		*envll;
 }	t_data;
