+++ includes/token.h
@@ -3,17 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   token.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: walnaimi <walnaimi@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: fdessoy- <fdessoy-@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/02 11:35:39 by lstorey           #+#    #+#             */
-/*   Updated: 2024/08/14 16:28:24 by walnaimi         ###   ########.fr       */
+/*   Updated: 2024/08/09 15:00:48 by fdessoy-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef TOKEN_H
 # define TOKEN_H
 
-# define MAX_ARG_STR 10240
 # include "minishell.h"
 # include <stdbool.h>
 typedef struct s_data t_data;
@@ -75,8 +74,7 @@ typedef struct s_token
  *****************************************/
 int		line_tokenization(t_data *data);
 int		chunky_checker(char *token, t_token *current_token, t_data *data);
-int 	check_and_handle_echo(t_token *current_token, t_token **prev_token, const char *delimiters, t_data *data);
-void 	free_tokens(t_token *head);
+int		check_and_handle_echo(t_token *current_token, t_token **prev_token, const char *delimiters, t_data *data);
 
 /*****************************************
  * src/parse/helper.c
@@ -90,16 +88,8 @@ int		search_token_type(t_token *token, t_type type);
  * src/parse/expand_env.c
  *****************************************/
 char 	*expand_env_variable(const char *input, size_t *i, t_data *data, size_t *new_len);
+void 	copy_env_value(char *result, const char *env_value, size_t *j);
 char 	*expand_env_variables(const char *input, t_data *data);
-
-/****************************************
- * src/parse/expand_env_utils.c
- ****************************************/
-void 	setup_env_variables(const char *input,t_data *data);
-void 	single_q(const char *input, t_data *data, char *res, size_t *i, size_t *j);
-void 	double_q(const char *input, t_data *data, char *res, size_t *i, size_t *j);
-void 	handle_env_variable(const char *input, size_t *i, t_data *data, char *result, size_t *j);
-void 	copy_env_value(char *result, const char *env_value, size_t *j, t_data *data);
 
 /*****************************************
  * in src/parse/ft_strtok.c
@@ -107,31 +97,18 @@ void 	copy_env_value(char *result, const char *env_value, size_t *j, t_data *dat
 char	*ft_strtok(char *str, const char *delim, t_data *data, t_token *cur_tok);
 int		ft_charinstr(char c, const char *str);
 int		ft_strcmp(char *s1, char *s2);
-
-/*****************************************
- * in src/parse/ft_strtok_utils.c
- *****************************************/
-char 	*remove_quotes(const char *str, t_data *data);
-char 	*skip_starting_delim(const char *str, const char *delim, char **target);
+char	*remove_quotes(const char *str, t_data *data);
 
 /*****************************************
  * in src/parse/echo.c
  *****************************************/
 void 	echoing(t_token *current_token, t_token **prev_token, const char *delimiters, t_data *data);
 char 	*concatenate_echo_args(t_token *current_token, const char *delimiters, t_data *data);
-char 	*validate_and_process_token(const char *target, t_data *data);
-void 	process_quoting_and_delimiters(const char *target, const char *delim, t_data *data, t_token *cur_tok);
-void 	handle_quote(const char *target, t_data *data, t_token *cur_tok);
 
 /*****************************************
  * in src/parse/init_token.c
  *****************************************/
 t_token	*init_token();
-
-/*****************************************
- * in src/parse/modify_str.c
- *****************************************/
-void 	modify_str(char *str);
 
 /*****************************************
  * in src/parse/chunky_check.c
@@ -148,7 +125,6 @@ int		ft_argument_check(char *token, t_token *current_token);
 int		check_builtin(char *token, t_token *current_token, t_data *data);
 int		check_echo_flag(char *token, t_token *current_token, t_data *data);
 int		check_flag(char *token, t_token *current_token, t_data *data);
-
 
 /***************************************
  * in src/parse/chunky_check_2.c
