+++ src/main.c
@@ -3,29 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: walnaimi <walnaimi@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: fdessoy- <fdessoy-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/13 10:12:51 by fdessoy-          #+#    #+#             */
-/*   Updated: 2024/08/12 18:19:41 by walnaimi         ###   ########.fr       */
+/*   Updated: 2024/08/05 10:23:59 by fdessoy-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	free_gang(t_data *data)
-{
-	if(data->vtoken != NULL)
-	{
-		free(data->vtoken);
-		data->vtoken = NULL;
-	}
-	if(data->ctoken != NULL)
-	{
-		free(data->ctoken);
-		data->ctoken = NULL;
-	}
-	free_tokens(data->token);
-}
 
 int	main(int argc, char **argv, char **env)
 {
@@ -46,24 +31,16 @@ int	main(int argc, char **argv, char **env)
 			status = sniff_line(data);
 			if (status == NULL_LINE)
 			{
-				/*free_gang(data);
-				if(data)
-					free(data);*/
 				printf("exit\n");
 				break ;
 			}
 			else if (status == 963)
-			{
 				continue;
-			}
 			else
-			{
 				execution(data, &env_ll);
-				//free_gang(data);
-			}
+			free_token(data->token);
+			data->token = NULL;
 		}
-		free_gang(data);
-		free(data);
 	}
 	else
 		ft_putstr_fd(ERR_ARG, 2);
